@@ -3,7 +3,7 @@ description: Consulta tus calificaciones de la UTVT
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -31,24 +31,23 @@ dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  simple_animations: ^4.0.1
-  http: ^0.13.4
+  simple_animations: ^5.0.2
+  http: ^1.1.0
   html: ^0.15.0
-  flutter_secure_storage: ^5.0.2
+  flutter_secure_storage: ^9.2.2
   change_app_package_name: ^1.0.0
-  admob_flutter: ^2.0.0
-  photo_view: ^0.13.0
-  share_plus: ^3.0.4
+  # admob_flutter: ^3.0.0
+  photo_view: ^0.15.0
+  share_plus: ^7.2.2
   path_provider: ^2.0.8
-  flutter_svg: ^1.0.0
+  flutter_svg: ^2.0.10+1
   contained_tab_bar_view: ^0.8.0
   provider: ^6.0.2
-  shimmer: ^2.0.0
-  font_awesome_flutter: ^9.2.0
+  shimmer: ^3.0.0
+  font_awesome_flutter: ^10.7.0
   url_launcher: ^6.0.20
 
 dev_dependencies:
@@ -60,8 +59,8 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
-  flutter_launcher_icons: "^0.9.2"
+  flutter_lints: ^4.0.0
+  flutter_launcher_icons: ^0.13.1
 
 flutter_icons:
   android: "launcher_icon"
@@ -73,7 +72,6 @@ flutter_icons:
 
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
@@ -81,7 +79,7 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-  #   - images/a_dot_burr.jpeg
+    #   - images/a_dot_burr.jpeg
     - assets/images/cuervo_perfil.png
     - assets/svg/
 
